+++ app/Http/Controllers/EspecialidadesController.php
@@ -22,6 +22,9 @@ class EspecialidadesController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('ModuloEspecialidades'))
+            abort(403, 'Permiso Denegado.');
+
         $especialidades = Especialidad::paginate(10);
         return view('especialidades.index', ['especialidades'=>$especialidades]);
     }
@@ -33,6 +36,8 @@ class EspecialidadesController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('ModuloEspecialidades'))
+            abort(403, 'Permiso Denegado.');
         return view ('especialidades.create');
     }
 
@@ -87,6 +92,9 @@ class EspecialidadesController extends Controller
      */
     public function edit($id)
     {
+        if (!Auth::user()->can('EditarEspecialidades'))
+            abort(403, 'Permiso Denegado.');
+
         $especialidad = Especialidad::findOrFail($id);
         return view('especialidades.edit', ['especialidad'=>$especialidad]);
     }
@@ -131,6 +139,8 @@ class EspecialidadesController extends Controller
      */
     public function destroy($id)
     {
+        if (!Auth::user()->can('EliminarEspecialidades'))
+            abort(403, 'Permiso Denegado.');
 
         Especialidad::destroy($id);
         return redirect('/especialidades')->with('mensaje', 'Especialidad eliminada satisfactoriamente');
